requirements.txt
# Local RAG System Dependencies

# Core RAG & ML
lancedb>=0.3.1
sentence-transformers>=2.2.2
transformers>=4.35.0
torch>=2.0.0
numpy>=1.24.0

# Document processing
pypdf>=3.17.0
PyMuPDF>=1.23.0  # More memory-efficient than pypdf
python-docx>=1.1.0
docx2txt>=0.8

# Text processing & chunking
tiktoken>=0.5.1
nltk>=3.8.1

# FastAPI & web
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
jinja2>=3.1.2
python-multipart>=0.0.6
aiofiles>=23.2.0

# HTTP client for LM Studio
openai>=1.3.0
httpx>=0.25.0

# Utilities
pydantic>=2.5.0
python-dotenv>=1.0.0
tqdm>=4.66.0
pyarrow>=14.0.0

# System monitoring
psutil>=5.9.0
watchdog>=3.0.0

# 2026 Upgrade: Advanced RAG Features
rank-bm25>=0.2.2  # BM25 keyword search for hybrid retrieval
ragas>=0.1.0  # RAG evaluation metrics
datasets>=2.14.0  # Required by RAGAS

# Optional: For better performance
# accelerate>=0.24.0  # For faster model loading
